Initial starterkit creation (#1)
This is the initial commit of code for the XM Cloud ASP.NET Core Starter
Kit. Some notes about it's contents:

- This is a replica of the new structure that has been approved for the
Next.js starter kit (foundation head).
- The repo favours a FED-First development approach as per the new
direction for XM Cloud. The container based approach is still included
here, and it's assets can be found in the `local-containers` folder.
- The components included are a direct port of the same components found
in the Next.js starter, replicating the logic and generated markup/css.
- There are currently 2 components that haven't been ported over as
we're blocked waiting for changes to the ASP.NET Core SDK (Navigation
and FEaaS). Both of these are called out in the main README to ensure
readers are aware.
- The main README contains initial steps to run the repo locally against
an XMC preview endpoint, using VS Dev Tunnels to connect Pages to the
local developer machine. Work will need to be done to refine this
process in the future but this will do for an initial beta release, so
we can garner initial feedback.
- This is currently using the "old" method of connecting to the XMC
Preview and Edge endpoints, it needs updating to support the ContextID
approach, but this is blocked until the SDK is supports this approach.