Improved Font loading (#19)
Currently fonts are loaded in a way, that it blocks the first paint of
the website. This drains the lighthouse score a bit. To avoid that, the
font loading was changed

<!-- Provide a general summary of your changes in the Title above -->
<!-- Apply the label "bug" or "enhancement" as applicable. -->

## Description / Motivation
<!-- Describe your changes in detail -->
Font loading is removed from the CSS and put into the .cshtml file as
link reference. That way fonts are now loading non-blocking in the
background
<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->
Closes https://github.com/Sitecore/xmcloud-starter-dotnet/issues/18

After applying the changes, fonts are still loaded correctly and
lighthouse does not complain anymore about render blocking resources
(fonts)

![image](https://github.com/user-attachments/assets/3341545d-47d1-43e3-842a-0a38a943fd66)


## Terms
<!-- Place an X in the [] to check. -->

<!-- The Code of Conduct helps create a safe space for everyone. We
require that everyone agrees to it. -->
- [x] I agree to follow this project's [Code of
Conduct](CODE_OF_CONDUCT.md).